op-acceptance-tests: Separate load tests from interop package
The acceptance tests dramatically increase Go test CI time to ~15 minutes per run. This is causing a lot of contention for CI resources, and delaying PR merges. This PR separates the load tests from the acceptance tests to reduce CI time.
