+++ op-acceptance-tests/tests/interop_loadtest/eoa_pool.go
@@ -1,4 +1,4 @@
-package loadtest
+package interop_loadtest
 
 import (
 	"sync"
