+++ op-acceptance-tests/tests/interop_loadtest/interop_load_test.go
@@ -1,4 +1,4 @@
-package loadtest
+package interop_loadtest
 
 import (
 	"math"
