+++ op-acceptance-tests/tests/interop_loadtest/metrics_collector.go
@@ -1,4 +1,4 @@
-package loadtest
+package interop_loadtest
 
 import (
 	"context"
